Update README.md
EventHub is a web application designed to help users easily create, manage, and view events. It allows users to register and login securely, and perform full CRUD operations on events such as creating new events, editing existing ones, deleting, and viewing the complete list of events. The application features a simple and clean user interface with responsive forms and tables, along with flash messages for success and error notifications to enhance user experience. Built using Django on the backend and HTML/CSS for the frontend, EventHub follows best practices for project structure and is ideal as a main project to showcase in a resume or portfolio.